Add a test program that imports RPM metadata into the database.
This certainly isn't shippable code right now.  The database's name is
hardcoded - it's called test.db.  It doesn't like it if you try to
import the same RPM twice (though, it might be okay - I haven't checked
the results).  It only does the main tables, and only most of the data
that goes in those.

But it does work.  Just pass RPMs on the command line and watch them go.
For checking what happened, I find it easiest to then do this after the
import is complete:

$ sqlite3 test.db
> .output BLAH
> .dump