Support importing multiple RPMs that came out of the same SRPM.
Basically, there should only ever be one project and source entry for a
single SRPM no matter how many RPMs it produces.  The builds table is a
little more questionable, and maybe I'm doing it wrong.  Anyway, each
RPM we import gets a key/value entry with the package's name, and then
all the files for that RPM get associated with that package.

I've left some demo code in loadRPM to dump the file list for each
import, as well as a filesInPackage function that implements the query
to figure this out.