Fix importing the same RPM twice.
All the work I did to support multiple RPMs from a single SRPM broke
what happens if you try to import the exact same RPM twice.  It was
allowing that which meant duplicate copies of all the file lists and
everything.  The new buildImported function checks to see if an RPM with
the same SHA1Header has already been imported and if so, we skip it.