Remove duplication in the key_val table.
When two builds of the same RPM are imported, they will likely share a
lot of duplicate information.  In the key_val table, the likely
duplicates are parts of NEVRA as well as most of the rpm-provides.  Get
rid of duplicates by searching for them first and returning the ID if a
key/value pair already exists.