Start adding version info to requires and provides.
I'm just putting these in as expressions represented by strings for the
moment.  Everything besides Requires: is still unsupported.  This is not
very well tested, especially not when it comes to querying the database
and seeing what happens.  I have done a quick import and looked at the
results, and everything seems to be getting imported fine though.