Add a new depclose program.
This doesn't do any dependency solving.  It only queries the database
for everthing that is potentially a dependency of a given RPM and
returns that list.  This is mostly proof of concept code for figuring
out if our database has everything we need in it and if we can do all
this quickly.