Handle multilib possibilities.
Basically, just return the packages for all architectures and let the
caller figure out which is appropriate.  This fits in with our approach
of not making decisions here.

This required changing getNEVRAsForGroupName, findGroupId, and
getRequirements to handle and return multiple values.
getValueForGroupName was also changed to take in a group ID instead of a
group name so it can be called across the results of the other
functions.

Finally, I added a ton of debugging prints to closeRPM inside
conditionals.  To enable these prints, you need to do the following:

$ cabal configure -f debug
$ cabal build

To remove them again, you just configure without "-f debug" and rebuild.