Update the database code to work with a db generated from schema.sql.
Changes were made to schema.sql recently to add indices to a bunch of
tables.  This is run by entrypoint.sh.  Haskell doesn't know how to do
the indices with persistent (you could use raw SQL, I think).
Unfortunately, persistent is also getting very confused by the file
types table and fails to migrate.  The result is that you can't
automatically generate the database with sqlite and then import with
haskell, unless you're using an older database and things are tricked
into not needing to run.

So, I killed all the migration and initDB code from Haskell.  You have
to run sqlite with schema.sql to generate a blank database before
importing.  I also added a definition for the FileType table to DB.hs
and code around that to make sure a reference to the right file type
ends up in the Files table.

I fixed a typo in schema.sql.

Finally, I am now storing the complete RPM package name in the Groups
table.  Storing just the Name tag results in failing constraints when
trying to import multilib versions of packages.