Move group name back to just being a name, not a NEVRA.
I think making this change in the first place was done without thinking
it through enough, and it caused problems for the UI.  The only reason I
did it in the first place is because the constraint in schema.sql
started failing.  And that only started failing because I just recently
started using schema.sql to initialize the database.

So in summary, this change should get things back to exactly as they
were working before schema.sql started being used.  Whether that state
is broken or not, we'll have to figure out.