Split rpm provide/conflict/obsolete into name and version.
This adds a nullable, unindexed column to key_val for additional data
associated with the key/val that will not normally be used in a where
clause. For the RPM provide, conflict, and obsolete fields, this column
is used to store the full expression while the "val_value" column is
used to store only the name, so that a provider for a particular string
can be more easily found. So something like

   Provides: libsomething >= 47.0-1

will be stored as

   ('rpm-provide', 'libsomething', 'libsomething >= 47.0-1')