Rearrange dependency gathering.
When gathering the providers for a given thing, convert straight from
group ids to Propositions instead of returning group names. This way
unnecessary requirements are not pulled in, and the provides Proposition
is built from the rpm-provide key_val information instead of from the
requirement expression.

Also combine the rpm-provide ane filename cases in closeRPM, because
sometimes a requires expression that looks like a filename is also
provided via rpm-provide.