Remove the mddb every time entrypoint.sh runs.
This is so we pick up changes in the database schema that may have been
committed.  Yes, it will slow things way down.  However, it's faster
than waiting for someone to manually go in and remove the database and
regenerate it.  While we are still in development, this seems like a
reasonable thing to do.  Once we're not changing the schema with some
frequency, we can probably revert this.