Import the Order tag from RPMs.
We will need this in the future when we are done with dependency solving
and need to start ordering the results for installation.  It is not
needed for dependency solving itself since using this tag appears to
also generate a Requires tag too.

If package A has "OrderWithRequires: B", that means B needs to be
installed before A.  In the database, I am using the key
"rpm-install-after" on a group to indicate it must be installed after
whatever is the value in the row.