Allow importing from a remote repo.
This is done by providing import with the path to the primary.xml.gz
file in repo metadata.  This code assumes that the metadata is
compressed, that it is stored in a repodata/ directory, and that the
path to the packages is formed by stripping the repodata/ part off
first.