Add the beginnings of support for loading yum groups into the mddb.
This commit quickly grew out of control.  I needed to rearrange
everything to take into account the fact that we support importing from
multiple sources.  Then I did some function renaming so that everything
in Import/ has the same kinds of names (which means you'll need to do
qualified imports).  And then I could add the new yum groups code.

Note that this doesn't currently do anything with the data it loads from
the comps.xml file.  I don't yet know how we are going to store this
information in the database, and I think a lot of the stuff in BDCS/ is
going to have to get rearranged to be less RPM-centric too.

So basically, don't use the new comps code right now.  It's mainly here
to (1) force me to think about how everything else should be rearranged,
and (2) to make sure it doesn't get lost.