Begin making the BDCS/ directory less RPM-centric.
The approach I'm taking is splitting each file up into an RPM-specific
part and a generic part.  The RPM part knows how to take a package (as a
[Tag]) and build a record from it.  The generic part knows how to insert
that record into the database and make sure it doesn't already exist.

I've only done the Sources.hs file in this commit because that one was
very simple.  More to follow.