Add a new MissingRPMTag exception.
DBException is a very general exception type.  Does it describe a
problem with adding stuff to the database?  With pulling things out
of the database?  Something else?  This doesn't make writing tests easy.

So, add this new more specific exception that is raised only when we try
to grab a specific tag out of a [Tag] and it does not exist.  More
exception types to follow as needs arise.