Add the beginnings of a test suite.
These are pretty lame right now - all I'm doing is attempting to run the
various mk* functions with a list of Tags and seeing if they raise the
exceptions they are supposed to.  I am not yet trying to construct
records with questionable data and seeing if we get it right.  The most
important thing here, however, is the structure.

The general strategy for test case writing is that the stuff in
BDCS/RPM/ should be the focus.  We don't need to test if RPMs were
parsed into [Tag] correctly, as that's the job of the haskell-rpm
module.  And we don't really need to test the code in BDCS/ since that
is basically just dumping a record into the database.  There's not much
code required to do that, and it's largely just testing that persistent
works.