Wrap processThing and all it calls in a read-only state.
I'm working on adding ostree support, which is going to require passing
a repo object all around.  With two parameters having to be passed to
everything, it makes sense to convert them into a state record.

The URL/file path being processed does not go into the state.  We could
be processing many RPMs so that piece of data will change for each.