Update db.cabal.
I had to update my local system, which means now I've got base-4.9.0.0
installed.  Update the cabal file to allow that.  And then with that
done, I had to remove the restriction on esqueleto, but it seems to be
building fine.

And finally, put a version restriction on the rpm module.  This is
needed to make sure we don't pull in the rpm module in hackage, which is
really not very useful.  In the future, I plan to make our haskell-rpm
module an official thing and get rid of the subtree.  So this is just
temporary.