Add a new CS.hs module.
This is for manipulating the content store - here, an ostree repo.  It
stores the contents of the RPMs we are importing, and returns the
checksums for each one.  We can then store these checksums in the mddb
and use that to know which files to pull out of the content store when
something is requested.

Note that there's a lot around this that does not work very well so far.