Pipe the RPM payload into the content store.
And pipe the checksums into the mddb.  One big note on this is that the
content store is not yet 100% correct.  For some reason, directories
don't have checksums.  Ownerships don't appear to be correct, and I
haven't checked that permissions are always right.  If you commit a
second package to the content store, the first one is no longer listed
in "ostree log".  There's probably more.