Install happy via cabal instead of via dnf.
If use dnf, I get the Fedora 25 package which is based on ghc-7.10.3 but
none of those libraries exist in the image.  I instead need to use the
cabal version, which is based on ghc-8.  And then since it gets
installed into the home directory, I need to set $PATH too.