Convert String to T.Text wherever possible.
Haskell also has too many string types.  In general, people suggest you
should use Text instead of String because it can be faster and because
it knows how to handle non-ASCII data.  The Data.Text module also
defines a variety of little functions that make manipulating them
easier, which you have to invent yourself if you are dealing with a list
of Chars.

There's likely a bunch of places where types are still being converted
back and forth between String and Text, and those should be dealt with
as they are discovered.  There's also a variety of places where we are
using ByteStrings.  As long as those are limited to files and network
streams, it's alright.

The next step in this will be to make haskell-rpm return Text for
everything.