Add an export program.
This adds a new function and some types to BDCS.CS for retrieving data
from a content store given a checksum from the mddb. There is also a new
function in BDCS.Files to fetch the Files records for a given group id,
returned as a conduit.

Export does not currently handle UID or GID, since those are complicated
and would require it to be run as root anyway, nor does it handle
xattrs, since we don't actually have any of those yet. It will probably
also blow up on anything that isn't a file, directory or symlink.