Fix ungzipIfCompressed in the uncompressed case.
For esoteric reasons that I do not understand, there can be no identity
conduit and therefore "awaitForever yield" does not work in the middle
of a pipeline.  But "mapC id" does, even though it feels more lame.

Also I've changed the imports to just pull everything from the Conduit
module, which is supposed to be our one-stop-shop for conduit
functionality.  It also gives all the list-related functions helpful C
suffixes to ensure they don't conflict with prelude.