Add importing from a dnf/yum repo URL
If the argument to import is not recognizable as something else, look
for repodata at <arg>/repodata/repomd.xml and import the primary.xml and
comps data contained therein. If the argument turns out not to be a repo
it will fail with an error from http-conduit or xml-conduit and probably
look real confusing, so, uh, FIXME I guess.

Added the start of an idea of throwing exceptions from import, which is
thrown if repomd.xml is readable and parseable but doesn't contain what
it's supposed to.