Add a new module for functions related to depsolving.
This adds a new type for logical propositions, a function for converting
formulas in said type to conjunctive normal form, and a function for
sat-solving a CNF formula.

depclose or something similar still needs to be modified to use the
Formula type, and at that point we'll have something resembling a real
dependency solver.