Rework how tools are built using cabal.
The problem was that the build section for each tool required listing
all the dependencies that tool was indirectly using through the library.
That is, if the library used esqueleto but the tool did not directly us
it, esqueleto still needed to be listed as a dependency.

The fix is to move all the tools into their own directory.  Then, I can
remove the redundant stuff from each tool's build section.