Add a Utils.Monad.hs file.
And put concatMapM in there.  We're using this in several places
already, and I am about to add one more.

Feel free to add other functions to this file, or other files to this
directory.  My idea is to keep the functions small and the files focused
on one thing each.