Handle errors when trying to export content to a tar file.
The primary errors that could occur come from the tar module itself - if
paths are too long or contain characters that are not supported by the
standard, that'd be an error.  I suppose we  could also see errors
coming from trying to read the database or write the filesystem.

On an error, create a message that will get displayed, abort the whole
thing, and erase any partially constructed tar file.