Only process one thing per invocation of runSqlite in export.
For some reason, when we start to process the second thing, we get the
following unhelpful error message:

export: SQLite3 returned ErrorMisuse while attempting to perform bind int64.

I can't find any information on where that comes from, and the sqlite
docs just say you are trying to use the library incorrectly.  Whatever.
Reworking the loops so that runSqlite is on the inside fixes things.