Begin adding support for passing export a list of RPM filenames.
This is instead of passing just a name in - the web API provides us with
a list of filenames, and we need to convert that into something that the
database knows about.  The filename can optionally include ".rpm" on the
end.  It should be smart enough to trim that off if it exists.

It's got a couple problems, though.

First the splitFilename function needs tests and also needs to live
somewhere else.

Second the query in nvraToGroupId needs to be fixed.  I've left the raw
sql query in a comment.  That needs to be turned into Haskell.

Third nvraToGroupId needs to support epochs and then be renamed to
nevraToGroupId.