Rearrange export to use more conduits.
The only difference needed between tar and directory exports is how
content store objects are written to disk, so rearrange the processThing
functions so that both export types share all of the code except for a
conduit sink.

With this, tar exports use the same logic as directory exports for mddb
entries with no checksums, so tar exports now ignore ghost files instead
of bombing out.

The placement the SQL monads is kind of gross since we still haven't
figured out the issue with exporting multiple things with one runSqlite
causing an ErrorMisuse.