Add the ability to associate various labels with files.
These are stored in the KeyVal table as a distinct type.  We could just
use strings, but I prefer the type safety of making them a datatype.  It
could prevent bugs in the future.  Multiple labels can be associated
with the same file.

For now I've only added two - Docs (anything in /usr/share/doc) and
Libraries (any .so or .a files).  There's really not much code involved
to support them.  You just add a new file with a matches function and
add it to the list in FileLabels.hs.  These existing tests are easy to
modify if we feel the need.

It would be more natural to call these "tags", but we already have that
name from the stuff in the RPM header.  Having two tag concepts for
totally different things and distinguishing between the two with
qualified imports would be confusing.