Fix exporting to a directory as a non-root user
A lot of directories in the content store are not user-writable, and
when we make one of these directories and try to write files to it,
that's an error. Instead of making the directories with the appropriate
permissions as we see them, save them up for a second pass.

This also adds a warning when exporting a something that does have a
parent directory owned by anything in the export. Some of these will be
files missing from a recipe, but some of them are because, for instance,
/lib is a symlink and not a directory, and we're not completely dealing
with that yet.