Makefile: make `make importer` grab `export` binary
This should make `make importer` copy out the `export` binary in
addition to the `import` binary, so you can play around with the tools
locally if you so desire.

It also makes the `importer` rule depend on Makefile.build, which
doesn't really do anything in this case since `importer` is PHONY and
will rebuild no matter what, but it's kind of best practice to do that
anyway.