Add a tool to look at what's in the database.
Command line handling is sloppy at the moment, but the basic idea is
there's subcommands that can take arguments.  Current commands are
limited to printing out groups and nevras and listing all files.

Feel free to add more stuff, but the majority of the code should live in
the library.