Add a "-l" option to the ls command.
The addition of a second option will require real option processing,
which I am still avoiding doing.

For now this just printing out user, group, and mtime.  Permissions and
size are hiding in the content store which I will probably have to write
a lot of weird other code for.