Add object type, mode, size, and symlink target to "ls -l" output.
Also, use printf to format the output columns better and rearrange the
code a bit to make it easier to add other types of things in the future.
This also includes a patch from dshea to make it build that should have
been a separate patch, but is too tangled up with the rest now to break
apart.