Fix ghost file output in the "ls -l" command.
Regular ls output was printing out more files than verbose output.  The
reason is that getMetadata was returning Nothing for ghost files,
because there is no checksum for ghost files.  And then the Nothings
were getting filtered out by CL.catMaybes.

Instead, always put the filename in the tuple returned by getMetadata
and use a Maybe value for the metadata object.  For ghost files, print
out a size of 0 and a mode that includes "ghost".