Convert the inspect subcommands into standalone programs.
The idea here is that the inspect program is a front end that will
eventually find the content store and mddb and pass those as arguments
to the subcommand so the user doesn't have to know all that.  For now it
just passes arguments through.  The subcommand must exist in $PATH.

The standalone subcommand programs can be run directly themselves, too.

More clever handling of subcommands and errors in inspect.hs is coming
in a later patch.