Add a lot of custom code to Setup.hs for subcommands.
(1) Make sure anything that starts with "inspect-" gets installed to
$libexecdir.  These are subcommands that are not intended to be run
directly by the user.  cabal-2.x allows us to do this much more simply,
but we're not using that version yet.  When we are, all this can go.

(2) Make sure $libexecdir ends with "weldr/".  This gets around not
being able to set it in the package-level cabal.config file.   This
might be taken care of in cabal-2.x, but I can't be certain.