Add key/val output to the ls subcommand.
These can be output when returning verbose or regular output.

Unfortunately, it also got a little mixed up with a code reorganization
patch required to make it easier to add another output option.  Instead
of building tuples as I go through the conduit, I've added a record
where most of the members can be Nothing.  At each step in the conduit,
we see whether various pieces of output were requested.  If so, get the
data and put it into the record.  At the end, examine that record to
display to the user.