Add json output to the ls command.
There's nothing really tricky about this.  Some functions moved out to a
more global scope so I could access them from both the verbose and json
printers.  Note that json output can be requested independently of other
options - you can get json output that is verbose or not, and with
keyvals or not.