Change how keyval lists are converted into json.
(1) The exact same code was in both the groups and ls commands.
Condense it down into one place, in the BDCS library.

(2) Change how LabelKeys are represented in json.  Instead of each
ending up as a key with a value of "true" in an object, create a list
named "labels" where each label just ends up being represented as text
in that list.  This seems cleaner to me.