mkLabel should return a Maybe.
Once more involved mkLabel functions start getting added, it could be
possible for them to fail.  Thus they should return a Maybe.  The apply
function needs to be adapted to handle this case.

While I was at it, I turned the apply function inside out.  It now
iterates over all the files once, and the various checks for each file.