Fix the import program to use content-store repos.
First, use the runCsMonad function where applicable.  This gets rid of
needing to use runExcepT, and it gets rid of the resource part of
runConduitRes (which means we can just use runConduit).  Second, use
mkContentStore to both create the store if it doesn't exist and open it
if it does.  Third, fix a filename mismatch between what cpio gives and
what's in the RPM headers.