Get the NPM module working with the new content store code.
This commit kind of grew and grew until it was too big to straighten
out.  It started by me wanting to rebase to get the latest database
schema changes befoer adding my own.  Once I had done that, all the NPM
code was broken compared to the new content store so it didn't build.

Fixing that required first fixing filesToObjectsC and all the rest of
the export code, which required straightening out how directories and
symlinks would be handled.  In the end, it was all too tangled up to
break into smaller, more understandable commits.