Switch back to using checkAndRunSqlite.
I was getting compile errors about functional dependencies between the
two MonadError types at work - CsError and String.  The trick is to
fully run the the ExceptT monad where we are doing the database stuff,
then process the result of that, and map the result into the result type
of the surrounding code.

Here, the mapping takes the form of converting it into the generic
CsError type.