Rename the importer directory to src.
The importer name comes from basically the very beginning of this repo,
when I was just working on an import tool and didn't know where it was
going.  Now, that directory holds everything and should just be named
src.

This patch just does the move - it doesn't fix up any references to
importer in files.