Modify various files to refer to the src/ directory.
One note here is that in a lot of the docker-related stuff, I'm calling
the directory where things happen "bdcs" instead of "src".  The latter
seems entirely too generic a term.  So it's "src" when it's the source
directory in the git repo, and "bdcs" when it's the directory docker
uses.