Update to use the new version of content-store.
On the plus side, this fixes some pretty big bugs that we were bound to
hit at some point.  On the middle side, it means our stores will now be
compressed which might require changes to our testing setup.  On the
minus side, it requires s/MonadIO/MonadResource/ in several places.