Rename the inspect subcommands.
First, the directory goes from Commands to subcommands.  Second, don't
include that directory in other-modules when building inspect.  Third,
lower case the names of the files.

These three actions are meant to show that these files are not part of
any library, but are in fact standalone programs that just happen to get
installed into a weird location and called indirectly through another
program.