Don't use ostree checkout when testing exports. Fix #116
since 29c5a11 we're using a new version of content-store which,
as clumens confirms, is not a valid ostree repo anymore! So modify
the tests to not use ostree checkout anymore.

Instead I use

    rpm -Uhv --root ... --nodeps --noscripts *.rpm

to install the RPMs into a directory and compare this directory
with the results of the export command!