Get rid of all uses of the head and tail functions.
These are partial functions, which means they are not defined for empty
or infinite lists.  Each use could potentially be a bug.  Most of the
time, this can be fixed with a case statement and some pattern matching
instead.

This also resulted in a couple of other changes:

* getDbVersion now throws an error if there's no user_version.

* There's a BadName exception now, in case we can't make sense of an
RPM's name.

* parseCompsGroups can now return a Nothing, which means extractGroups
will throw those out.  None of this code is currently being used, but
this is more correct behavior anyway.