Add a module for building virtual filesystem trees
Build a tree of the filesystem metadata before performing an export.
This fixes some issues regarding package ordering, export file ordering,
and symlinks to directories. See http://weldr.io/Filesystem-requirements/
for more detailed information on the problems that this solves.

Since this fixes the requirement that symlinks to directories exist
before children of these directories are created, remove the requirement
that filesystem be included first (or at all) in an export.