Handle some new warnings from new ghc
ghc-8.2 has a new warning about unlisted modules. List the ones we
hadn't been listing for exectuables, and turn off the warning (and turn
off the warning about unknown warnings, so it works cleanly with older
ghc) when compiling tests, so that we can continue to implicitly
recompile the library.