Compile and test bdcs during docker runtime
this commit merges the build+unit test stage with the integration
test stage so that testing will work again.

Because we mount the local directory to the running
docker container all build/test operations must now happen during
runtime (NOTE: -v mounts read-only during docker build). Without
this change the build stage installs bdcs inside a filesystem layer
which is not accessible to the integration-test container and it
fails.