The export function no longer takes an mddb argument.
Instead it is now set up to operate from inside runSqlite, or
checkAndRunSqlite, or runSqlPool (which is what bdcs-api needs).  Errors
are returned via throwError, and error handling cleanup has been
removed.  It's now expected that the caller will clean up on error.

The standalone export tool was modified as needed.