Fix the case of a symlink in the way of a file
Sometimes finding a symlink in the way of a path being added means that
there's a symlink in the way of a path being added. Fix this case so
that, instead of following the symlink and recursing, it either compares
two symlinks for equality, follows the symlink to move placeholder directories
around, or throws an error.