Update various build requirements.
First, update codec-rpm and content-store to their latest versions.
This is to keep the build requirements among all these modules in sync.

Second, update exceptions to allow the latest versions too.  Various
modules are requiring the new version of exceptions, and it doesn't
break the bdcs build at all, so allow that.

Finally, restrict how recent a version of monad-logger is allowed.
Newer versions pull in a newer version of resourcet that results in
build failures.  This version of resourcet also conflicts with the
version of conduit we are currently using.  Don't update the
monad-logger restrictions without testing it out first.