Add a callProcessLogged function that does what it says.
This replaces most calls to the callProcess function out of
System.Process.  This version also logs how it was called, its stdout,
and any stderr.  If the command failed, an exception is raised as well.

Note that there's one call not being logged at the moment, because its
types are pretty rough.