Add total number of results to getGroupsLike
In order to support pagination we need to know the total number of
results, before applying offset and limit, so do a 2nd SQL query for
that number and return it along with the results.

This should still be faster than returning all results and doing offset
and limit in Haskell.