Use ExportType throughout the library.
This moves the file extension checking up out of the BDCS library and
into the export command - we make the check there, set the proper expor
type, and then had that off to the library for all further comparisions.

A more thorough patch would also add a command line argument to the
export program so you can give any filename you want for any export
type, but command line argument processing is not fun.