Add an option to the export command to specify the export type.
This breaks apart the existing link between the name of the export
artifact and the type of the artifact.  You can now give any name you
want to any type you want.

This changes the command line interface for the export command.  You
must now give "-t <some type>".  Because the number of things you need
to provide on the command line continues to grow, I've also made it so
the destination is given via "-d <dest>".  Finally, error handling on
command line options should be better.