Switch to using the typed-process module.
There doesn't appear to be a process module package in Fedora, but there
is a typed-process package.  The resulting code has a little more lift
and MonadBaseControl in it which is annoying, but the module on a whole
provides a lot of nice functions.  I especially like using exceptions in
callProcessLogged instead of a case statement branch that does nothing.