Remove script support from the database.
We're not using this for anything, and we haven't since it was added so
clearly it's not too important.  I've left the ReqType stuff in - it
still seems useful to know whether a requirement is there because it's
needed by the thing you're installing, or because of a script, or what.