Only use cabal-2.x for building bdcs.
This has one big advantage:  "scope: private" is now supported for
executables, which means they end up getting installed into libexecdir.
We don't have to do that ourselves, which means all that stuff in
Setup.hs can go away and it becomes the default simple version.

It also requires the autogen-modules stuff getting added, but that's not
a big deal.