Drop Dockerfile and related files
currently we use Docker only for testing purposes and the files
intended to be used for packaging the software to be used by users
are not really in use.  Remove these since clumens has suggested
more simplistic approach based on bdcs.rpm.

Also build & test on the latest Fedora image from upstream!